Refactor to avoid violating sealed packages.
  Move flow and contract/state classes into their own packages.
