+++ cordapp-contracts-states/src/main/kotlin/com/template/states/TemplateState.kt
@@ -1,0 +1,12 @@
+package com.template.states
+
+import com.template.contracts.TemplateContract
+import net.corda.core.contracts.BelongsToContract
+import net.corda.core.contracts.ContractState
+import net.corda.core.identity.AbstractParty
+
+// *********
+// * State *
+// *********
+@BelongsToContract(TemplateContract::class)
+data class TemplateState(val data: String, override val participants: List<AbstractParty> = listOf()) : ContractState
